fix redux actions file name for login and register
